requirements.txt
flask>=3.0,<4
flask-restful>=0.3.10,<1
flask-migrate>=4.0,<5
flask-sqlalchemy>=3.1,<4
flask-marshmallow>=1.0,<2
flask-jwt-extended>=4.6,<5
flask-cors>=4.0,<5
marshmallow>=3.20,<4
marshmallow-sqlalchemy>=1.0,<2
python-dotenv>=1.0,<2
passlib>=1.7,<2
apispec[yaml]>=6.3,<7
apispec-webframeworks>=1.0,<2
psycopg2-binary>=2.9,<3
gunicorn>=21.2,<23
